requirements.txt
torch>=2.1.0
diffusers>=0.25.0
transformers>=4.36.0
Pillow>=10.0.0
opencv-python>=4.8.0
scipy>=1.11.0
numpy>=1.24.0
tqdm>=4.66.0
huggingface-hub>=0.20.0
timm>=0.9.12
einops>=0.7.0
omegaconf>=2.3.0
pytest>=7.4.0
